+++ src/__deprecated__/navbar/index.js
@@ -1,3 +1,3 @@
 // @flow
-export { NavbarV2 } from "./navbar";
+export { Navbar } from "./navbar";
 export { BaseNavbar } from "./components/base-navbar";